ENH: Cleanup of NoInterrupt.
When used with ipykernel, our suspension of signals fails since it
restores the default handler.  We get arround this by replacing the
pre_ and post_handler_hook methods to prevent this.  See:

https://github.com/ipython/ipykernel/issues/328

The previous version tried hijacking signals.signal, but this failed,
probably becuase the original version was retrieved before we could
hijack this.

Still needs testing with the ipykernel package.